backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==0.21.1
httpx>=0.18.1,<0.19.0
xrpl-py==1.7.0
xumm-sdk-py==1.0.3
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0
aiosqlite==0.19.0
websockets>=9.0.1,<11.0 